Fixes a typo causing caps to display contained items when not worn (#4689)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

A small oopsie made it so caps still displayed contained items when they
were held and not worn. This fixes that.

Thanks to joooks for reporting

# Explain why it's good for the game
Consistency


# Testing Photographs and Procedure
spawn in as SO, with mgoggles, switch between hand and equipped


# Changelog
:cl:
fix: Fixed caps displaying contained items on character even without
being worn.
/:cl: