Removes Stunprod from LV522 (#4712)
# About the pull request

Removes stunprod from LV522 Chances Claim

# Explain why it's good for the game

I've been getting DM'd about this stunprod for over a week and since the
synth council doesn't believe a synth stunning xenos on the frontline to
be combat synthing I'm just going to remove it


Edit for the sake of a better explanation. Realistically this item
should have never been on LV522 I had placed it before not knowing its
fully power and I'll take responsibility for that no map should have an
item with the power to stun every xeno on it regardless if it only has 6
charges or not

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:SpartanBobby
maptweak: Removes stunprod from LV522 
/:cl: