Fix a source of bad icon operations (#4720)
# About the pull request

This PR updates the stat panel examine of human mobs to use a dummy copy
of the examined mob to fix runtimes related to a bad icon operation from
a workaround in icon2html for humans. It is not a perfect copy to limit
the upper end of cost for rendering humans, but should be fairly
accurate visually and forced to face south. Initial testing seems to
report 0 to 0.2 seconds per mob.

This PR doesn't eliminate the possibility of an icon2html of a human,
but I am unsure if its wise to just remove that workaround. Even if it
was just removed for this situation, the resulting icon would be empty
for the human mob.

# Explain why it's good for the game

Fixes an issue that can result in tens of thousands of runtimes.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![image](https://github.com/cmss13-devs/cmss13/assets/76988376/8f715e79-f4f2-48c7-b59a-1a6ec9e71d6d)

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/fce430b1-9c76-42ea-ba80-447d30a99651)

</details>


# Changelog
:cl: Drathek
fix: Fixed a cause of bad icon operations
/:cl: