Fixes deleted mobs not unbuckling (#4717)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

This makes mobs unbuckle on deletion.

Why this matters:
 - Leftover references as buckled_mob on the buckled object
- The object cannot be used by someone else, since someone is still
buckled to it
- In most extreme case eg. shuttle chairs, the shuttle will pull the
deleted mob out of nullspace and back to chair when traveling

# Explain why it's good for the game
Less bugs and less confused admins having to re-spawn an entire tank
over it

# Testing Photographs and Procedure
Tested buckling/deleting without errors