Renames UPP survivor synth from combat to support synth (#4703)
# About the pull request

Renames UPP Trijent Nightmare synthvivor from UPP Combat Synth to UPP
Support Synthetic.
It wasn't a combat synth in the first place.

# Explain why it's good for the game

Its not a combat synth and this round that just ended by time of
creation I saw a friendly synth get killed and cremated because its ID
said combat synthetic, so its obviously causing problems; and I'm going
out on a limb and going to guess these problems were not intended.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
spellcheck: UPP Synth Survivor on the Trijent Dam nightmare is now
called a Support Synthetic instead of a Combat Synthetic, as its not a
combat synthetic
/:cl: