Adds splints to the survival pouch & updates its description (#4724)
# About the pull request

Adds splints to the survival pouch that survivors spawn with, increases
the survival pouch storage space from 6 to 7 & updates the survival
pouch description.

# Explain why it's good for the game

Generally, unless you're actively inside of a survivor hold, getting a
bone-break is essentially a death-sentence, as if you manage to kill
whatever threat is attacking you, you are forced to either stop moving,
or run the risk of gaining IB for each step you take. This wouldn't be
an issue if it weren't for the large amount of RNG in getting a
bone-break. You could just as easily have a ravager maul you to death
and not break any bones as you could get a broken bone within one or two
slashes from a runner.

The current description is out-dated and straight-up wrong. The
description is based off of the old survival pouch, which could hold
metal sheets.


# Testing Photographs and Procedure
It does indeed work

</details>


# Changelog

:cl: foxtrot1322
add: Added splints to the survival pouch
add: Increased survival pouch storage space by 1
spellcheck: Updates the survival pouch's description
/:cl: