Add Non-Hostile UPP ERT (#4713)
# About the pull request

Adds a non-hostile UPP ERT alternative to the hostile ERT spawn. Has the
same chance to spawn as the hostile version.

# Explain why it's good for the game

In lore the UPP enjoy a cold war esque relationship with the UA, a
possibly neutral ERT helps better reflect this.

Furthermore, as more UPP are added to the game, the need to help shift
the community meta regarding UPP away from 100% hostile to 50/50 hostile
is going to become more important, this would be 1 small step.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: UPP ERT's have a chance to be neutral to the USCM.
/:cl:

---------

Co-authored-by: Steelpoint <alexander.henley@hotmai.com>
Co-authored-by: harryob <me@harryob.live>