Hotfixes Megaphone X Offset (#4741)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Insufficient testing made me once more commit this mistake, because the
langchat image is not regenerated the text keeps offsetting every time,
drifting away and offscreen.

This hotfixes it.

closes #4742 


# Changelog
:cl:
fix: Re-fixed Megaphone above-head-chat drifting to the left.
/:cl: