Fixes OB Warheads deletions (#4740)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

`runtime error: addtimer called with a callback assigned to a qdeleted
object. In the future such timers will not be supported and may refuse
to run or run with a 0 wait`

Makes OB Warheads and especially cluster delete timely (after their
explosion) saving us incertainties for most warheads, and a few hundreds
of runtimes for cluster.

# Explain why it's good for the game
Cleaner logs, i can't see what's going on in there geez

# Testing Photographs and Procedure
Fired a cluster checking runtimes. That's it.


# Changelog

No player facing changes