Disables code in icon2html that causes bad icon operations (#4732)
# About the pull request

This PR disables MrStonedOne's workaround in icon2html specifically for
humans. Rather than throwing a bad icon operation, the resulting icon
will just be blank.

# Explain why it's good for the game
Situations that cause tens of thousands of runtimes in a round need to
be eliminated.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

1. Inject Carbon and Nitrogen into self
2. Inject Oxygen into self
3. Observe a reaction for a blank icon (you) rather than throwing
errors.

![image](https://github.com/cmss13-devs/cmss13/assets/76988376/efbd7eea-69c5-4968-8c24-8f79c66970df)

</details>


# Changelog
:cl: Drathek
fix: Disabled code in icon2html that is causing bad icon operations
/:cl: