M39 being able to use vertical grip again [BUFF] (#4627)
# About the pull request

M39 receives a buff, which allows it to use a attachment that make sit
viable.

# Explain why it's good for the game

M39 as it stands is quite useless, the vertical grip might help it
become a viable choice over the M4RA, the MK2 and the M37 :)


# Testing Photographs and Procedure

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl: stalkerino
balance: m39 is able to use vertigrip
/:cl: