Fixes Observer Death Messages being delayed (#4775)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Long standing bug - due to waiting for not one but two full database
entities loading, the death message in dchat is delayed.

It's also symptomatic of a scheduling problem with DBQM as you can see
deaths occuring in "batches" every couple seconds, indicating the
queries are not processed timely - but this is out of scope for this PR.

Instead it just moves the observer logging to take place before DB ops.

# Explain why it's good for the game
Feels less clunky.

# Testing Photographs and Procedure
Just spawned stuff and abombed it.


# Changelog
:cl:
fix: deadchat death messages should now display immediately rather than
being delayed a couple seconds.
/:cl: