4 New Haircuts (#4728)
# About the pull request

## Some new haicuts

Looking at the current sheet it occurred to me to add a few current
popular hairstyles that have passed the community evaluation.

It's a bit hard for me to write text and descriptions due to my poor
language skills, but I'll try to describe as much as I can.

# Explain why it's good for the game

First of all, playing on the server we try to associate ourselves with
the character, try to be in his skin. His appearance plays an important
role. A couple of beautiful hairstyles will please the eye of players
and add a little more variety

# Photographs
<details>
<summary>Screenshots & Videos</summary>

## Gentle Ponytail


![image](https://github.com/cmss13-devs/cmss13/assets/82751385/340409c0-28ac-494b-935e-b21eed177468)


![hair_gentleponytail](https://github.com/cmss13-devs/cmss13/assets/82751385/f58b4fe0-3bb4-41be-b016-77edfa26b22e)

## Edgar haircut


![image](https://github.com/cmss13-devs/cmss13/assets/82751385/69020adb-bc87-4f68-9f56-e3334bf5b741)


![hair_edgar](https://github.com/cmss13-devs/cmss13/assets/82751385/c0477be4-f10c-4622-a496-b3d405631df3)

## Emo Bun


![image](https://github.com/cmss13-devs/cmss13/assets/82751385/605352c2-c35c-49da-bbc1-2dbf472fde44)


![hair_emobun](https://github.com/cmss13-devs/cmss13/assets/82751385/93432acf-2116-403e-9cc7-7ffd772fb6b0)

## Taper haircut


![image](https://github.com/cmss13-devs/cmss13/assets/82751385/18378922-f72b-4060-b66d-b3aae4e16a23)


![hair_taper](https://github.com/cmss13-devs/cmss13/assets/82751385/15e72936-c981-4e0e-9834-ac0f4740c605)

</details>


# Changelog

:cl: XDinka
add: Added four new haircuts: gentle ponytail, edgar haircut, emo bun,
taper haircut.
/:cl: