Adds ~50 tips (#4772)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Adds 38-or-something more tips, with some information on JTAC and
Settings esp for new players.

# Explain why it's good for the game
Nobody seems to bother doing it and the tips get stale over time.


# Changelog
:cl:
add: Added about 50 new tips.
/:cl:

---------

Co-authored-by: ihatethisengine <115417687+ihatethisengine@users.noreply.github.com>
Co-authored-by: forest2001 <41653574+realforest2001@users.noreply.github.com>
Co-authored-by: Zonespace <41448081+Zonespace27@users.noreply.github.com>