Fixes custom ERT calling (#4783)
# About the pull request
Removes announce_dispatch_message variable as it was a defunct duplicate
of a different variable.
Adds and fixes the announcement settings for custom ERTs.

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
fix: Fixes custom ERT calling broadcasting when it should not.
code: Removes unused vars from ERT procs.
code: Renames the announce var in ERT procs to be more indicative of
what it does.
add: Added a setting on custom ERTs for announcing beacon was received.
/:cl: