Remove Direct Hit Limb Penalty against Xenos (#4788)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

## Simple version

Selecting a body part on doll at bottom right applies an inherent
accuracy debuff to shots (an hand is harder to hit than chest).

This penalty still applied to sprite-click attacks on Xenomorphs,
despite the localized damage having no effect.

This means that you'd have to constantly reset your aimed part to chest
every time you try to shoot a Xeno, or would suffer a "free" penalty.

## Detailled version

Old flow went as follow:
 * Get base Hit Chance and Roll
* If it's only a near hit, or target's not the shot clicked target, the
shot targeted limb is reset to a random one
 * Hit chance receives a penalty based on aimed limb
 * We proceed checking hit chance vs roll

Because of the above problem this is changed to:

* Get base Hit Chance and Roll
* If it's a near hit (including limb penalty), a near xeno hit
(excluding limb penalty), or not the sprite-clicked target, aimed limb
is reset to a random one
* Else if the target is a xeno and it was a proper direct hit, there is
no penalty -- functionally similar to a chest shot
* Else Finally, apply per-limb penalty as usual
* We proceed checking hit chance vs roll

# Explain why it's good for the game
Avoids an invisible noobtrap debuff for just about everyone in the game
that isn't in the know - and the neccessity to constantly reset
targeting to chest for medics.

# Testing Photographs and Procedure
The problem was tested by a player on discord -- the fix i tested just
was spawning stuff and shooting up the place observing RNG misses. I did
not take detailed statistics over time to confirm.


# Changelog
:cl:
fix: Sprite-click shots onto Xenos are no longer affected by
limb-targeting penalty, because it was an accuracy debuff when there is
no inherent benefit to targeting Xeno limbs.
/:cl: