shrinks UPP SMG mag sprite size (#4813)
# About the pull request

When I was making the UPP gun sprites, apparently I was taken over by
the shitsprite demon and as a consequence the proportions on this
magazine are completely off compared to other magazines in the game and
their ammo amount. This PR fixes this grave error.

# Explain why it's good for the game

A magazine holding 35 rounds shouldn't look bigger than equivalent
magazines that hold more ammunition.


# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
imagedel: shrinks the type-19 SMG stick magazine to a more reasonable
size
/:cl: