Fixes GC deletion issues around reagents (#4833)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

Fixes a potential GC consistency issue around reagents and
reagents_containers

In theory this should do nothing because `reagents` is *supposed* to get
soft GC'ed anyway when it is swapped out.

There is something more to it somewhere, because on live bottles get
deleted and retain their `reagents`. In this case it's a circular
reference preventing deletion.

# Explain why it's good for the game
backend consistency / long term performance