Fixes being able to load a flamer tank with no flamer unit (#4825)
# About the pull request
Closes https://github.com/cmss13-devs/cmss13/issues/4824
Makes it so that you have to have the Broiler equipped in your back
before you are allowed to switch out fuel tanks.
If you used the verb after unequipping the broiler from your back and
then dropped it, it was possible to remove the tank from the flamer by
clicking it.
This caused issues because the tank would then teleport to you from
anywhere after switching fuel types again.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Bugfix
# Changelog
:cl:
fix: Fixes being able to create ghost tanks from a Broiler-T unit.
/:cl:

---------

Co-authored-by: Drathek <76988376+Drulikar@users.noreply.github.com>