Fix job entry message wiki links (#4823)
# About the pull request

This PR is a followup to the issues created by #4449 fixing job
entry_message_body links to the wiki.

/datum/controller/configuration now has a variable `is_loaded` that can
be checked to see if configuration has already occurred such as the case
for the ASO where subsequent creations of the job datum were not hitting
the proc that would be called on the config load signal.

# Explain why it's good for the game

Broken wiki links harm the new player experience.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


![aso](https://github.com/cmss13-devs/cmss13/assets/76988376/73c0b8e4-78f6-4f34-bfbe-9f74b7f05b14)

![CO](https://github.com/cmss13-devs/cmss13/assets/76988376/30b627bc-afe3-4064-b745-fecab5925970)

![marine](https://github.com/cmss13-devs/cmss13/assets/76988376/25250d85-445f-4506-8d91-3c12d500d20c)

</details>


# Changelog
:cl: Drathek
fix: Fixed various job's entry messages having broken links to the wiki
/:cl: