Nightmare Survivor Hostility Indicator (#4628)
# About the pull request
Adds a big bold message for Nightmare survivors to know if they're
hostile to marines or not.
This does not apply to natural map-specific survivors, only those in
nightmare inserts.
<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

# Explain why it's good for the game
Makes it clearer for nightmare inserts to indicate if survivors are
hostile. Good example of this is UPP on trijent are non hostile but this
isn't always clear.
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Added a big line to Nightmare survivor spawns that indicates
hostility.
/:cl: