Reimplements halloween pumpkin helmets event (#4623)
# About the pull request

<!-- Remove this text and explain what the purpose of your PR is.

Mention if you have tested your changes. If you changed a map, make sure
you used the mapmerge tool.
If this is an Issue Correction, you can type "Fixes Issue #169420" to
link the PR to the corresponding Issue number #169420.

Remember: something that is self-evident to you might not be to others.
Explain your rationale fully, even if you feel it goes without saying.
-->

This reimplements the Halloween Pumpkins spawning on map and being
carvable as helmets which was apparetnly a thing a number of years ago

# Explain why it's good for the game
Variety

# Changelog
:cl: Firartix and Frans_Feiffer
add: Readded Halloween pumpkin helmets.
/:cl:

---------

Co-authored-by: harryob <me@harryob.live>