Xeno Tunnel and Shuttle Building/Anchoring Changes (#4831)
# About the pull request

This PR removes the ability to build walls (and other ground structures)
on top of xeno tunnels, and removes the ability to anchor wall girders
on shuttle tiles and tunnels. Barricades can still be moved on top of
the tunnel's tile and anchored, though they can't be built while
standing on it.

# Explain why it's good for the game

Being able to build walls on top of a gaping hole in the ground is
pretty dumb, and results in weird interactions like having to right
click to use the tunnel that's currently being covered up by a wall. The
shuttle tiles were changed because, while you can't complete the girders
if they're in a shuttle, you can still anchor them down. That doesn't
seem intended since you can't build them while in a shuttle, and you
also can't build/anchor barricades in shuttles.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>


</details>


# Changelog
:cl:IowaPotatoFarmer
del: Removed the ability to build ground structures or anchor wall
girders on top of xeno tunnels.
fix: Fixed wall girders being anchorable on shuttle tiles.
/:cl: