Adds icon for hive core and cluster to Xeno tacmap (#4820)
There was supposed to be an icon for the hive core but it doesn't seem
to exist.

I added one for hive core and cluster


![dreammaker_2023-11-01_21-26-47](https://github.com/cmss13-devs/cmss13/assets/25027759/38f4d39f-4e1d-4874-bc8d-1a9a020a4604)


![2023-11-01_21-18-05](https://github.com/cmss13-devs/cmss13/assets/25027759/e0cf5302-004b-4a2a-9370-a8fcfac961a6)

My sprites even at this level are awful. If someone who is actually
artistic could make better ones I will use those. Likewise if you would
like to make 2 states for the core, damaged and healthy I can restore
the intended functionality of displaying a different icon for a damaged
core.

<!-- Remove this text and explain what the purpose of your PR is.


# Explain why it's good for the game
# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Put screenshots and videos here with an empty line between the
screenshots and the `<details>` tags.

</details>


# Changelog
:cl:
add: Xeno tacmap icons for hive core and clusters.
/:cl:

---------

Co-authored-by: Lee B <lee.b@josephgiles.com>