Fix req ammo vendor not dispensing to the table (#4856)
# About the pull request

This PR fixes #4852 caused by #4803 simply because the vendor is not
adjacent to the table it is supposed to vend to.

# Explain why it's good for the game

It was unintentional for this to change.

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>



https://github.com/cmss13-devs/cmss13/assets/76988376/e56607a3-7641-4c91-b585-fddeefb97098

After changing from turf to get_turf(user) per Fira:



https://github.com/cmss13-devs/cmss13/assets/76988376/a7a55df4-531c-40ea-a8a3-9b28d550f656

</details>


# Changelog
:cl: Drathek
fix: Fix the ColMarTechAutomated Munition Vendor in req not vending to
the table.
/:cl: